feat: class interface
