+++ components/client/typescript/src/schemas/index.ts
@@ -1,7 +1,8 @@
 import { Static, Type } from '@sinclair/typebox';
 import { StacksEvent } from './stacks';
 import { BitcoinEvent } from './bitcoin';
-import { IfThisSchema } from './predicate';
+import { BitcoinIfThisSchema } from './bitcoin/if_this';
+import { StacksIfThisSchema } from './stacks/if_this';
 
 const EventArray = Type.Union([Type.Array(StacksEvent), Type.Array(BitcoinEvent)]);
 
@@ -10,7 +11,7 @@ export const PayloadSchema = Type.Object({
   rollback: EventArray,
   chainhook: Type.Object({
     uuid: Type.String(),
-    predicate: IfThisSchema,
+    predicate: Type.Union([BitcoinIfThisSchema, StacksIfThisSchema]),
   }),
 });
 export type Payload = Static<typeof PayloadSchema>;
